consolidate spinner components and shared types across the frontend
