+++ frontend/app/(auth)/login/page.tsx
@@ -9,6 +9,7 @@ import client from "../../api/client";
 import AuthCard from "../../components/auth/AuthCard";
 import AlertBanner from "../../components/auth/AlertBanner";
 import FormField from "../../components/auth/FormField";
+import { ButtonSpinner } from "../../components/Spinner";
 import { getErrorMessage } from "../../lib/getErrorMessage";
 
 export default function LoginPage() {
@@ -130,7 +131,7 @@ export default function LoginPage() {
                 >
                     {loading ? (
                         <div className="flex items-center gap-2">
-                            <div className="w-4 h-4 border-2 border-white/30 border-t-white rounded-full animate-spin" />
+                            <ButtonSpinner />
                             {t("signingIn")}
                         </div>
                     ) : (
