+++ frontend/app/(auth)/reset-password/page.tsx
@@ -7,6 +7,7 @@ import { motion, AnimatePresence } from "framer-motion";
 import { Lock, ArrowRight, CheckCircle2, AlertCircle, Eye, EyeOff } from "lucide-react";
 import client from "../../api/client";
 import { useTranslations } from "next-intl";
+import { ButtonSpinner } from "../../components/Spinner";
 import { getErrorMessage } from "../../lib/getErrorMessage";
 
 export default function ResetPasswordPage() {
@@ -260,7 +261,7 @@ export default function ResetPasswordPage() {
                 >
                     {loading ? (
                         <>
-                            <div className="w-4 h-4 border-2 border-white/30 border-t-white rounded-full animate-spin" />
+                            <ButtonSpinner />
                             {t("form.resetting")}
                         </>
                     ) : (
